requirements.txt
Django==3.2.2
Pillow==8.2.0
djangorestframework==3.12.4
